move buffer.h and allocator.h into memory folder
