+++ src/memory/buffer.h
