+++ src/rpc.h
@@ -7,10 +7,10 @@
 
 #pragma once
 
-#include "allocator.h"
 #include "async.h"
-#include "buffer.h"
 #include "function.h"
+#include "memory/allocator.h"
+#include "memory/buffer.h"
 #include "serialization.h"
 #include "synchronization.h"
